reel comments

+++ app/Http/Controllers/Reel/Traits/ReelComments.php
@@ -1,11 +1,12 @@
 <?php
 namespace App\Http\Controllers\Reel\Traits;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Http\Request;
 use App\Models\Reel;
 use App\Models\ReelComment;
+use App\Models\CommentHeart;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Resources\ReelsComments;
+use Illuminate\Support\Facades\Validator;
 
 trait ReelComments{
 
@@ -19,7 +20,7 @@ trait ReelComments{
     }
 
     protected function reelsCommentsDelete(Request $request, $reelId, $id) {
-        $del = ReelComment::where(['reel_id' => $reelId, 'id' => $id])->delete();
+        $del = ReelComment::where(['reel_id' => $reelId, 'id' => $id, 'user_id' => Auth::id()])->delete();
         if(!$del){
             return $this->failure('Comment is Not Exists.');
         }
@@ -37,13 +38,33 @@ trait ReelComments{
         if(empty($reel)){
             return $this->failure('Reel Not Found.');
         }
-        $user_id = Auth::id();
         $new = new ReelComment;
-        $new->user_id = $user_id;
+        $new->user_id = Auth::id();
         $new->reel_id = $reel->id;
         $new->comment = $request->comment;
         $new->save();
         return $this->success('Comment Added Successfully.');
+    }
+
+    public function CommentToggleHeart(Request $request, $reelId, $id) {
+        $reel = Reel::find($reelId);
+        if(empty($reel)){
+            return $this->failure('Reel Not Found.');
+        }
+        $add = CommentHeart::where(['reel_id' => $request->reel_id, 'id' => $id, 'user_id' => Auth::id()])->first();
+        if(!empty($add)){
+            $add->delete();
+        }else{
+            $add = new CommentHeart;
+            $add->reel_id = $request->reel_id;
+            $add->user_id = Auth::id();
+            $add->save();
+        }
+        $hearts_count = CommentHeart::where(['reel_id' => $request->reel_id, 'id' => $id ])->count();
+        return $this->success('Heart Toggled Successfully.', [
+            'love_count' => $hearts_count
+        ]);
+
     }
     
 }
